(fix) add working code for CaughtPokemonGallery and display in App
